Program Create Database MATLAB
The "ZCreateDatabase.mlapp" file was developed using MATLAB App Designer to create and manage a face database. This application detects faces, crops them, extracts embeddings using the FaceNet model, and then stores the embeddings and identity labels in a "data.mat" file, which serves as the primary database for the facial recognition system.