Program Face Recognition MATLAB
The "ZFaceRecognition.mlapp" file was developed using MATLAB App Designer as an application for performing the facial recognition process. This application takes facial images from the camera, detects faces using Haar Cascade, extracts embeddings through the FaceNet model, and then matches them with the embeddings in the "data.mat" file. The matching results are used to determine whether a face is recognized or not, while also sending control signals to the Arduino Uno according to the recognition results.